# docs/api-gateway.openapi.yaml
openapi: 3.0.3
info:
  title: API Gateway
  version: "1.0.0"
  description: |
    Прокси над Converter, Auth и ComfyUI.
servers:
  - url: http://localhost:3000
paths:
  /health/live:
    get:
      summary: Liveness probe
      responses:
        "200":
          description: Alive
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
  /health/ready:
    get:
      summary: Readiness probe
      responses:
        "200":
          description: Ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
  /health/startup:
    get:
      summary: Startup probe
      responses:
        "200":
          description: Started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /api/v1/:
    get:
      summary: API info
      responses:
        "200":
          description: Info
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  status:
                    type: string

  /api/v1/convert:
    post:
      summary: Convert SVG → JSON (proxy Converter)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: SVG файл
              required: [file]
      responses:
        "200":
          description: Scene JSON
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlannerScene"
        "400":
          description: Invalid input
        "502":
          description: Upstream error

  /api/v1/render:
    post:
      summary: Convert JSON → SVG (proxy Converter)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlannerScene"
      responses:
        "200":
          description: SVG string
          content:
            image/svg+xml:
              schema:
                type: string
        "400":
          description: Invalid input
        "502":
          description: Upstream error

  /api/v1/login:
    post:
      summary: Login (proxy Auth)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Token + user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"

  /api/v1/users/{id}:
    get:
      summary: User profile (proxy Auth)
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /api/v1/users/{id}/files:
    get:
      summary: List files (proxy Auth)
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: Files info
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true

  /api/v1/users/{id}/svg:
    get:
      summary: Get SVG (proxy Auth)
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: SVG
          content:
            image/svg+xml:
              schema:
                type: string
    post:
      summary: Upload SVG (proxy Auth)
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: SVG файл
              required: [file]
      responses:
        "200":
          description: Saved

  /api/v1/users/{id}/pdf:
    get:
      summary: Get PDF (proxy Auth)
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: PDF
          content:
            application/pdf:
              schema:
                type: string
    post:
      summary: Upload PDF (proxy Auth)
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: PDF файл
              required: [file]
      responses:
        "200":
          description: Saved

  /api/v1/users/{id}/pdf-files:
    get:
      summary: Get PDF by name (proxy Auth)
      parameters:
        - $ref: "#/components/parameters/UserId"
        - in: query
          name: name
          schema:
            type: string
          required: true
          description: PDF filename
      responses:
        "200":
          description: PDF
          content:
            application/pdf:
              schema:
                type: string
        "404":
          description: File not found

  /api/v1/users/{id}/png:
    get:
      summary: Get PNG (proxy Auth)
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: PNG
          content:
            image/png:
              schema:
                type: string
    post:
      summary: Upload PNG (proxy Auth)
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: PNG файл
              required: [file]
      responses:
        "200":
          description: Saved

  /api/v1/users/{id}/png-to-svg:
    post:
      summary: PNG → SVG passthrough (proxy Auth)
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: PNG файл
              required: [file]
      responses:
        "200":
          description: SVG
          content:
            image/svg+xml:
              schema:
                type: string

  /api/v1/users/{id}/png-to-json:
    post:
      summary: PNG → JSON via paired SVG (proxy Auth+Converter)
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: PNG файл
              required: [file]
      responses:
        "200":
          description: Scene JSON
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlannerScene"

  /api/v1/users/{id}/svg-edited:
    get:
      summary: Get edited SVG by name (proxy Auth)
      parameters:
        - $ref: "#/components/parameters/UserId"
        - in: query
          name: name
          schema:
            type: string
          required: true
      responses:
        "200":
          description: SVG
          content:
            image/svg+xml:
              schema:
                type: string
    post:
      summary: Upload edited SVG (proxy Auth)
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: SVG файл
              required: [file]
      responses:
        "200":
          description: Saved

  /api/v1/users/{id}/svg-json:
    get:
      summary: Convert original SVG to JSON by name (proxy Auth+Converter)
      parameters:
        - $ref: "#/components/parameters/UserId"
        - in: query
          name: name
          schema:
            type: string
          required: true
      responses:
        "200":
          description: Scene JSON
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlannerScene"

  /api/v1/users/{id}/svg-edited-json:
    get:
      summary: Convert edited SVG to JSON by name (proxy Auth+Converter)
      parameters:
        - $ref: "#/components/parameters/UserId"
        - in: query
          name: name
          schema:
            type: string
          required: true
      responses:
        "200":
          description: Scene JSON
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlannerScene"

  /api/v1/users/{id}/json:
    get:
      summary: Get JSON by name (proxy Auth)
      parameters:
        - $ref: "#/components/parameters/UserId"
        - in: query
          name: name
          schema:
            type: string
          required: true
      responses:
        "200":
          description: JSON
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
    post:
      summary: Upload JSON file (proxy Auth)
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: JSON файл
              required: [file]
      responses:
        "200":
          description: Saved

  /api/v1/users/{id}/json-edited:
    get:
      summary: Get edited JSON by name (proxy Auth)
      parameters:
        - $ref: "#/components/parameters/UserId"
        - in: query
          name: name
          schema:
            type: string
          required: true
      responses:
        "200":
          description: JSON
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
    post:
      summary: Upload edited JSON as raw data (proxy Auth)
      parameters:
        - $ref: "#/components/parameters/UserId"
        - in: query
          name: name
          schema:
            type: string
          required: false
          description: Имя файла (опционально, если не указан - генерируется автоматически)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              description: Raw JSON данные для сохранения
      responses:
        "201":
          description: Saved
          content:
            application/json:
              schema:
                type: object
                properties:
                  path:
                    type: string
                  filename:
                    type: string
        "400":
          description: Invalid JSON

  /api/v1/users/{id}/json-edited-pdf:
    post:
      summary: Save edited JSON and generate PDF report (proxy Auth+PDF)
      parameters:
        - $ref: "#/components/parameters/UserId"
        - in: query
          name: name
          required: true
          schema:
            type: string
          description: Имя файла (file_id) без расширения, используется для поиска оригинального JSON и сохранения edited JSON
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              description: Raw JSON данные редактированной планировки
      responses:
        "200":
          description: PDF report
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "400":
          description: Invalid JSON or missing name
        "404":
          description: Original JSON not found
        "502":
          description: PDF generation failed

  /api/v1/users/{id}/json-to-svg:
    post:
      summary: Render scene JSON to edited SVG (proxy Auth+Converter)
      parameters:
        - $ref: "#/components/parameters/UserId"
        - in: query
          name: name
          required: true
          schema:
            type: string
          description: Базовое имя исходного файла (без .svg добавится автоматически)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlannerScene"
      responses:
        "201":
          description: SVG saved to edited
          content:
            application/json:
              schema:
                type: object
                properties:
                  path:
                    type: string
                  filename:
                    type: string
        "400":
          description: Invalid input
        "502":
          description: Converter error

  /api/v1/comfyui/submit:
    post:
      summary: Submit ComfyUI job (proxy ComfyUI)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                gray_image:
                  type: string
                  format: binary
                  description: Серое изображение (UploadFile gray_image)
                reference_image:
                  type: string
                  format: binary
                  description: Референс-изображение (UploadFile reference_image)
                user_id:
                  type: string
                  description: Идентификатор пользователя
              required: [gray_image, reference_image, user_id]
      responses:
        "200":
          description: Submitted
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id:
                    type: string
                  message:
                    type: string
        "502":
          description: Upstream error

  /api/v1/comfyui/progress/{job_id}:
    get:
      summary: Get ComfyUI job progress (proxy ComfyUI)
      parameters:
        - in: path
          name: job_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Progress
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "404":
          description: Not found

  /api/v1/comfyui/download/{user_id}/{filename}:
    get:
      summary: Download ComfyUI file (proxy ComfyUI)
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
        - in: path
          name: filename
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File
          content:
            application/octet-stream:
              schema:
                type: string
        "404":
          description: Not found

components:
  parameters:
    UserId:
      in: path
      name: id
      required: true
      schema:
        type: string
  schemas:
    HealthResponse:
      type: object
      properties:
        status:
          type: string
      example:
        status: ready
    LoginRequest:
      type: object
      properties:
        login:
          type: string
        password:
          type: string
      required: [login, password]
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      properties:
        id:
          type: string
        login:
          type: string
        name:
          type: string
    PlannerScene:
      type: object
      description: React-planner scene JSON; структура передается как есть.
      additionalProperties: true